Normalize tabs for underlining output
If we don't do this, then we treat tabs as one character long and our
underlines mismatch the code. But we also can't assume that tabs are
4 or 8 characters long, since that depends on the output medium. So we
manually replace literal tabs with spaces. (This could be a problem if
there are tabs inside a string but I won't worry about it.)