Refactor: minimize matching token sequence distance by default
By default, the matcher finds a starting token, and then scans forward
for an ending token. This may result in a non-minimal range if there is
another starting token between the beginning and end of the range,
which would have to be accounted for by the caller.

For example, a match starting from '(' and going until ')' might match
this sequence of tokens:

    foo ( bar ( ) )
        ^^^^^^^^^

This isn't really desirable because we would rather have the tightest
set of open and close parentheses. This commit changes the default
behavior to instead match this subsequence:

    foo ( bar ( ) )
              ^^^