Refactor to pass filename to linter functions
The original interface directly passed the list of tokens as a
parameter:

    @linter.register
    def flag_something(tokens):
        ...

This is a fragile interface because as soon as we need to pass more
information into a linting function, we would have to modify the
function signature of every linting function!

An alternative to consider would be passing new information via keyword
arguments, which would require functions to register which information
they want to receive in a system like this:

    # Always have a **kwargs to soak up the keyword arguments are
    # passed but we don't want.
    @linter.register
    def flag_something_1(tokens, **kwargs):
        ...

    @linter.register
    def flag_something_1(filename, **kwargs):
        ...

But this looks a little bit awkward and it's unclear what the
ramifications are for this method in the long run.

The method used here is to wrap all of the information into a single
`SourceCode` object, which is passed as the first argument to every
function:

    @linter.register
    def flag_something(source):
        if source.tokens:
            ...

This also has the additional property of allowing us to attach helper
functions or properties to the passed-in object for convenience.