Bugfix: Include subpackages in the distribution
When trying to add the subpackage 'matcher' to the package 'lint381', it
mysteriously started being unable to import the subpackage when
installed via 'pip install', but not when installed 'pip install -e'.
The fix can be found here: http://stackoverflow.com/q/14553787/344643

The subpackages were not bundled with the package, meaning that only the
top-level source files were included (because I explicitly listed
'lint381' as a package). 'pip install -e' probably worked because I
imagine that it would be doing something like symlinking this directory
into the installed-packages directory, which would of course preserve
the directory structure.

The issue was debugged by manually going into the directory where the
packages were installed, and observing that only the top-level source
files were present for some reason.

Like an actual python, this language seems intent on murdering me by
means of painfully slow asphyxiation.