Bugfix: Display errors in the order that they appear
Previously, errors were displayed in the order that they are found,
which often happened to be the order that they appear in the source
code. But occasionally tokens were flagged in a different order than
they appeared in the source code, which is bad for the approach of
resolving errors first to last (as many people are used to doing with
regular compiler error messages).